Add a simple command line parser helper
The 'cmdline' module parses command line arguments and extracts the use
input.

This commit also includes a dos2unix conversion on the lexer file as
well as a small change to the 'lex' function (it now passes the tokens
along instead of printing them).

Future commits should be smaller than this, but this patch proved to be
a tricky one to edit.